solve router issue
